requirements.txt
# Core Frameworks & Server
fastapi==0.110.0
uvicorn==0.29.0

# Database
pymongo[srv]==4.7.0
motor==3.3.2  # Async MongoDB Driver

# Task Queue & Messaging
celery==5.3.6
redis==5.0.1  # For Celery backend
kafka-python==2.0.2  # Kafka for event streaming

# Data Processing & Utilities
python-dotenv==1.0.1  # Load environment variables
pydantic==2.6.4  # Data validation
pandas==2.2.1  # Data processing (if needed)
numpy==1.26.4  # Numeric operations
sqlalchemy==2.0.29  # If needed for SQL parsing
tenacity==8.2.3  # Retry logic
loguru==0.7.2  # Advanced logging

# Webhooks & Notifications
httpx==0.27.0  # Async HTTP requests
